adding gmap3.
not sure if we are going to use gmap3 int he long run but for getting
it running it is working well. also, changed the readmore link to not
load new pages via ajax. this is so that we don't have to put scripts
in the body or have a script that knows which page is current